+++ packages/social-controllers/src/index.test.ts
@@ -1,6 +1,10 @@
-import { SocialService, socialServiceName } from '.';
+import { SocialController, SocialService, socialServiceName } from '.';
 
 describe('social-controllers exports', () => {
+  it('exports SocialController class', () => {
+    expect(SocialController).toBeDefined();
+  });
+
   it('exports SocialService class', () => {
     expect(SocialService).toBeDefined();
   });